ExplicitNonNullaryApply sad story for Java methods
What we'd like is to leave any call to a method that was defined in Java
(or any "universal" method, like ##, eq, isInstanceOf) to be exempt.

It appears Scalameta/Semanticdb/Scalafix doesn't capture that
information, and thus basically all of them get parens... :(

Apparently in Metals they use the presentation compiler to recover all
the missing info and I know in ScalaClean they have their own nsc
Traverser, I'm guessing for the same reason.